Fix Update Bug
Updating a picture with a new file now correctly refreshes the page,
displaying the new image. Changed several things to fix this: (1) when
a file is uploaded for the first time, a timestamp is appended to its
filename on AWS and on the server; (2) when a picture is updated, the
old file is destroyed on AWS and a new one is created. (The name of the
new file is the same but with a different timestamp appended.); (3) the
picture object is then updated on the server.